i replace int height by size_t height

+++ 16-binary_tree_is_perfect.c
@@ -40,7 +40,7 @@ int height(const binary_tree_t *tree)
  * Return: height of the tree
  */
 
-int binary_tree_height(const binary_tree_t *tree)
+size_t binary_tree_height(const binary_tree_t *tree)
 {
 	if (tree == NULL)
 	{
